Move some errors to their own sub packages
